[CLS] Fix ‘jumping’ carousel
My previous usage of `content-visibility: auto;` was causing the page to
jump as the carousel neared or left the viewport. To solve this, I’ve
applied `content-visibility` to only subsequent (and therefore
off-screen) carousel items, meaning the first item always reserves the
correct space in the UI irrespective of whether it’s the current item.