Make Contact page’s Calendly async
As the Calendly widget on my contact page is above-the-fold, `defer`ring
it is a little too aggressive, especially as it being in-`body` means
that it’s queued behind SpeedCurve etc. Switching to `async` keeps it
off the Critical Path but it’ll hopefully run a little sooner.

Body (72 chars)