Switch Calendley `preconnect` to `dns-prefetch`
It turns out that `c839db0a` wasn’t a simple CORS thing (are they
ever?). Because `calendly.com` doesn’t response with an
`Access-Control-Allow-Origin` header, we can’t reuse the same
CORS-enabled connection that our `preconnect` opened. In fact, we can’t
use `preconnect` at all without forcing a wasted connection, a lost
round trip, and opening a final connection JIT and in-band. In this
case, `dns-prefetch` is actually faster as it avoids waste.