Drop Calendly `preconnect` down to `dns-prefetch`
I was wrong in 650efab7a562515bd26de600e0409a00d31ec953. It turns out
that the reason Chrome isn’t reusing the TCP/TLS from the `preconnect`
had nothing to do with CORS, but is down to the fact that requests made
with `Sec-Fetch-Mode: navigate` (e.g. `iframe`s) are placed into
a separate connection pool[1], meaning only their DNS can be reused.
A `preconnect` is actually quite wasteful in this scenario, and using
a humble `dns-prefetch` is the only way to get any real gains.

1. chromium.org/developers/design-documents/network-stack/preconnect